requirements.txt
django
google-generativeai
markdown2
gunicorn
whitenoise